requirements.txt
django
djangorestframework
requests
python-dotenv
pytest
django-background-tasks